+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Fix i18n string with dot 
+- When creating task in completed list, completedAt field is correctly set
 
 ### Removed